Fix merge

+++ map-backend/src/main/java/pl/pwr/bdmap/dto/NodeDTO.java
@@ -4,8 +4,8 @@ import java.sql.Timestamp;
 
 public record NodeDTO(
         int id,
-        int posX,
-        int posY,
+        double posX,
+        double posY,
         int blockedBy,
         Timestamp timestamp,
         String nodeType
